feat: adding jwt configs to be imported globally; creating routes and controllers for user; creating autorize jwt access

+++ src/services/user.service.js
@@ -6,7 +6,7 @@ dotenv.config()
 
 export const createUser = async (userData) => {
   const salt = process.env.SALT || 10
-  const [ password, ...newUser ] = userData  
+  const { password, confirmPassword,  ...newUser } = userData  
   const encryptPass = bcrypt.hashSync(password, salt)
   try{
     await prisma.user.create({
@@ -16,43 +16,66 @@ export const createUser = async (userData) => {
       }
     })
   }catch(err){
-    throw new ErrorHandler({statusCode: 400, message: err.message})
+    throw new ErrorHandler({statusCode: 409, message: "Email already exists"})
   }
 }
 
-export const deleteUser = async ({id}) => {
-  await prisma.user.delete({
-    where: {id},
-  })
+export const deleteUser = async ({id, req}) => {
+  if(req.auth.id  != id || !req.auth.type) {
+    throw new ErrorHandler({statusCode: 401, message: "You do not have permission to delete this user"})
+  }
+  try {
+    await prisma.user.delete({
+      where: {id},
+    })
+  }catch(err) {
+    throw new ErrorHandler({statusCode: 404, message: "User does not exist"})
+  }
+  
 }
 
-export const updateUser = async ({id, data}) => {
+export const updateUser = async ({id, data, req}) => {
+
+  if(req.auth.id  != id) {
+    throw new ErrorHandler({statusCode: 401, message: "You do not have permission to update this user"})
+  }
+  
   const update = await prisma.user.update({
     where:{
       id
     },
-    data
+    data: {
+      updated_at: new Date().toJSON(),
+      ...data
+    }
   })
-
+  
   if(!update?.email) throw new ErrorHandler({ statusCode : 404, message: "User not found"})
   delete update.password
-
+  
   return update
 }
 
 
 export const getUser = async ({id}) => {
+  
+  if (!id )  {
+    const users = await prisma.user.findMany()
+    users.forEach(user => delete user.password)
+
+    return users
+  } 
+
+
   const user = await prisma.user.findUnique({
-    where: {
-      id
-    }
-  })
+    where: {id: id},
+  }) 
 
-    if(!user?.email) throw new ErrorHandler({ statusCode: 404, message: "User not found"})
+  if(!user?.email) throw new ErrorHandler({ statusCode: 404, message: "User not found"})
 
-    
-    delete user.password
-    return user
+  
+  delete user.password
+  return user
   
 }
 
@@ -60,9 +83,10 @@ export const loginUser = async ({email, password}) => {
   const user = await prisma.user.findUnique({
     where: { email, }
   })
+
   if (!user?.email) throw new ErrorHandler({statusCode: 401, message: "Email or password incorrect"})
   
-  const match = bcrypt.compareSync(user.password, password)
+  const match = bcrypt.compareSync(password ,user.password)
   
   if (!match) throw new ErrorHandler({statusCode: 401, message: "Email or password incorrect"})
